moves resume routes to its own router
